separating StudentDashboard in many component + updating import in content.js
